requirements.txt
allure-pytest==2.2.1b1
allure-python-commons==2.2.1b1
attrs==17.2.0
backports.functools-lru-cache==1.4
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
crayons==0.1.2
EasyProcess==0.2.3
enum34==1.1.6
future==0.16.0
idna==2.6
pluggy==0.5.2
py==1.4.34
pytest==3.2.1
PyVirtualDisplay==0.2.1
requests==2.18.4
selene==1.0.0a11
selenium==2.53.1
six==1.10.0
urllib3==1.22
webdriver-manager==1.4.5
